Refactor project structure and update configurations
- Removed old gunicorn configuration file and added a new one in the configs directory.
- Deleted outdated test script and replaced it with a new implementation for testing ML predictions.
- Added model configuration file to define model descriptions and prediction methods.
- Created a new module for model utilities, including loading models and making predictions.
- Updated test cases for ML predictions and home page to ensure proper functionality.
- Added pytest configuration to suppress warnings during test runs.
