+++ .devcontainer/docker-compose.prod.yml
@@ -7,8 +7,8 @@ services:
       target: prod
     ports:
       - "90:90"
-    volumes:
-      - ..:/APP:ro,cached
+    # volumes:
+    #   - ..:/APP:ro,cached
     healthcheck:
       test: ["CMD-SHELL", "curl --fail http://localhost:90/healthz || exit 1"]
       interval: 10s     # how often to run the check
